# Conflicts:
#	lib/store/settings_store.dart
#	pubspec_base.yaml
#	res/values/strings_ar.arb
#	res/values/strings_bg.arb
#	res/values/strings_cs.arb
#	res/values/strings_de.arb
#	res/values/strings_en.arb
#	res/values/strings_es.arb
#	res/values/strings_fr.arb
#	res/values/strings_ha.arb
#	res/values/strings_hi.arb
#	res/values/strings_hr.arb
#	res/values/strings_id.arb
#	res/values/strings_it.arb
#	res/values/strings_ja.arb
#	res/values/strings_ko.arb
#	res/values/strings_my.arb
#	res/values/strings_nl.arb
#	res/values/strings_pl.arb
#	res/values/strings_pt.arb
#	res/values/strings_ru.arb
#	res/values/strings_th.arb
#	res/values/strings_tl.arb
#	res/values/strings_tr.arb
#	res/values/strings_uk.arb
#	res/values/strings_ur.arb
#	res/values/strings_yo.arb
#	res/values/strings_zh.arb

+++ lib/core/auth_service.dart
@@ -1,3 +1,5 @@
+import 'dart:io';
+
 import 'package:cake_wallet/core/totp_request_details.dart';
 import 'package:cake_wallet/routes.dart';
 import 'package:cake_wallet/src/screens/auth/auth_page.dart';
@@ -40,7 +42,9 @@ class AuthService with Store {
     final encodedPassword = encodedPinCode(pin: password);
     // secure storage has a weird bug on macOS, where overwriting a key doesn't work, unless
     // we delete what's there first:
-    await secureStorage.delete(key: key);
+    if (Platform.isMacOS) {
+      await secureStorage.delete(key: key);
+    }
     await secureStorage.write(key: key, value: encodedPassword);
   }
